moved internal interfaces to api
